fix: 날짜 구하는 메서드 수정

+++ src/main/java/com/fastcampus/reserve/interfaces/order/dto/response/OrderItemInfoResponse.java
@@ -10,7 +10,7 @@ public record OrderItemInfoResponse(
         String imageUrl,
         Integer maxGuestCount,
         Integer baseGuestCount,
-        int day,
+        long day,
         String checkInTime,
         LocalDate checkInDate,
         String checkOutTime,
